% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpb matrix.R
\name{pb_matrix}
\alias{pb_matrix}
\title{Gerar matriz Plackett Burman com FrF2}
\usage{
pb_matrix(
  nruns = 12,
  nfactors = 7,
  factor.names = c("A", "B", "C", "D", "E", "F", "G"),
  randomize = FALSE
)
}
\arguments{
\item{nruns}{Numero de ensaios, padrao 12.}

\item{nfactors}{Numero de fatores.}

\item{factor.names}{Nomes dos fatores.}

\item{randomize}{Se TRUE, randomiza a ordem.}
}
\value{
Objeto de planejamento do FrF2.
}
\description{
Gerar matriz Plackett Burman com FrF2
}
