% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rread clipboard dpb.R
\name{read_clipboard_dpb}
\alias{read_clipboard_dpb}
\title{Ler tabela do Excel via clipboard para Plackett Burman}
\usage{
read_clipboard_dpb(sep = "\\t")
}
\arguments{
\item{sep}{Separador, padrao tab.}
}
\value{
data.frame com colunas numericas.
}
\description{
Copie a tabela do Excel incluindo o cabecalho e use esta funcao.
}
